payment via paypal done, database updated, and notification model removed

+++ EZ-Wheels/GenericRepo/PaypalRepo.cs
@@ -5,7 +5,7 @@ namespace Car_Rental_APIs.GenericRepo
 {
     public interface PaypalRepo
     {
-        Task<Payment> CreatePaymentAsync(decimal amount, string currency);
+        Task<Payment> CreatePaymentAsync(float amount, string currency);
         Task<Payment> ExecutePaymentAsync(string paymentId, string payerId);
     }
 }
